Contact Page
The contact page has a field to accept the email address of a user, the Request Number changes the permission to allow the database to reveal the password, the next button get number will display the number of the user. Currently it isn't working as it is saying user is not found.